requirements.txt
asgiref==3.7.2
awsebcli==3.16.0
botocore==1.13.28
cement==2.8.2
certifi==2022.12.7
colorama==0.3.9
Django==3.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
future==0.16.0
idna==2.7
jmespath==0.9.4
matplotlib==3.2.1
mysqlclient==2.0.1
pathspec==0.5.9
pip==22.3.1
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2023.3
PyYAML==3.13
requests==2.20.1
semantic-version==2.5.0
setuptools==65.6.3
six==1.11.0
sqlparse==0.4.4
typing_extensions==4.7.1
wheel==0.38.4
wincertstore==0.2
